Переделал разметку для экрана создания заявки. Refactor

+++ app/src/main/java/com/example/a2023_q2_elmirov/data/network/model/LoanConditionsModel.kt
@@ -1,7 +1,7 @@
 package com.example.a2023_q2_elmirov.data.network.model
 
 data class LoanConditionsModel(
-    val maxAmount: Long,
+    val maxAmount: Int,
     val percent: Double,
     val period: Int,
 )
