+++ app/src/main/java/com/example/a2023_q2_elmirov/data/network/model/LoanRequestModel.kt
@@ -1,7 +1,7 @@
 package com.example.a2023_q2_elmirov.data.network.model
 
 data class LoanRequestModel(
-    val amount: Long,
+    val amount: Int,
     val firstName: String,
     val lastName: String,
     val percent: Double,
