+++ app/src/main/java/com/example/a2023_q2_elmirov/domain/entity/Loan.kt
@@ -3,7 +3,7 @@ package com.example.a2023_q2_elmirov.domain.entity
 import java.time.LocalDateTime
 
 data class Loan(
-    val amount: Long,
+    val amount: Int,
     val firstName: String,
     val lastName: String,
     val percent: Double,
