+++ app/src/main/java/com/example/a2023_q2_elmirov/domain/entity/LoanConditions.kt
@@ -1,7 +1,7 @@
 package com.example.a2023_q2_elmirov.domain.entity
 
 data class LoanConditions(
-    val maxAmount: Long,
+    val maxAmount: Int,
     val percent: Double,
     val period: Int,
 )
